Multi-attribute value function sensitivity analysis

MAVF_sensitivity computes summary statistics for multi-attribute value scores of x and y given a range of swing weights for each attribute

MAVF_sensitivity(data, x, y, x_wt_min, x_wt_max, y_wt_min, y_wt_max)

Arguments

data

A data frame

x

Variable from data frame to represent x attribute values

y

Variable from data frame to represent y attribute values

x_wt_min

Lower bound anchor point for x attribute swing weight

x_wt_max

Upper bound anchor point for x attribute swing weight

y_wt_min

Lower bound anchor point for y attribute swing weight

y_wt_max

Upper bound anchor point for y attribute swing weight

Value

A data frame with added variables consisting of sensitivity analysis summary statistics for each product or service (row).

Details

The sensitivity analysis performs a Monte Carlo simulation with 1000 trials for each product or service (row). Each trial randomly selects a weight from a uniform distribution between the lower and upper bound weight parameters and calculates the mult-attribute utility score. From these trials, summary statistics for each product or service (row) are calculated and reported for the final output.

See also

MAVF_score for computing the multi-attribute value score of x and y given their respective weights SAVF_score for computing the exponential single attribute value score

Examples

# Given the following data frame that contains \code{x} and \code{y} attribute
# values for each product or service contract, we can compute how the range of
# swing weights for each \code{x} and \code{y} attribute influences the multi-
# attribute value score.

df <- data.frame(contract = 1:10,
                 x_attribute = c(0.92, 0.79, 1.00, 0.39, 0.68, 0.55, 0.73, 0.76, 1.00, 0.74),
                 y_attribute = c(0.52, 0.19, 0.62, 1.00, 0.55, 0.52, 0.53, 0.46, 0.61, 0.84))

MAVF_sensitivity(df, x_attribute, y_attribute, .55, .75, .25, .45)

#>    contract x_attribute y_attribute MAVF_Min MAVF_1st_Q MAVF_Median MAVF_Mean
#> 1         1        0.92        0.52   0.7319     0.7572      0.7810    0.7805
#> 2         2        0.79        0.19   0.5122     0.5474      0.5815    0.5808
#> 3         3        1.00        0.62   0.8290     0.8476      0.8682    0.8675
#> 4         4        0.39        1.00   0.5426     0.5727      0.6031    0.6036
#> 5         5        0.68        0.55   0.5872     0.6196      0.6351    0.6349
#> 6         6        0.55        0.52   0.4909     0.5257      0.5399    0.5399
#> 7         7        0.73        0.53   0.6121     0.6433      0.6610    0.6604
#> 8         8        0.76        0.46   0.6035     0.6347      0.6560    0.6555
#> 9         9        1.00        0.61   0.8245     0.8436      0.8647    0.8640
#> 10       10        0.74        0.84   0.7424     0.7646      0.7748    0.7752
#>    MAVF_3rd_Q MAVF_Max MAVF_Range
#> 1      0.8030   0.8262     0.0943
#> 2      0.6133   0.6459     0.1337
#> 3      0.8868   0.9050     0.0760
#> 4      0.6329   0.6641     0.1215
#> 5      0.6505   0.6808     0.0936
#> 6      0.5550   0.5874     0.0965
#> 7      0.6778   0.7066     0.0945
#> 8      0.6761   0.7048     0.1013
#> 9      0.8838   0.9025     0.0780
#> 10     0.7860   0.8075     0.0651