Single attribute value function

SAVF_score computes the exponential single attribute value score of x

SAVF_score(x, x_low, x_high, rho)

Arguments

x

Numeric vector of values to score

x_low

Lower bound anchor point (can be different than min(x))

x_high

Upper bound anchor point (can be different than max(x))

rho

Exponential constant for the value function

Value

A vector of the same length as x with the exponential single attribute value scores

See also

SAVF_plot for plotting single attribute scores SAVF_preferred_rho for identifying the preferred rho

Examples

# The single attribute x is bounded between 1 and 5 and follows an exponential
# utility curve with rho = .653

x <- runif(10, 1, 5)
x

#>  [1] 2.472664 3.447078 2.658225 1.849568 3.930975 2.838340 1.398265 4.080634
#>  [9] 1.387139 1.103361

## [1] 2.964853 1.963182 1.223949 1.562025 4.381467 2.286030 3.071066
## [8] 4.470875 3.920913 4.314907

SAVF_score(x, x_low = 1, x_high = 5, rho = .653)

#>  [1] 0.6666629 0.8608644 0.7137408 0.4595172 0.9200178 0.7542928 0.2471371
#>  [8] 0.9348383 0.2410695 0.0704365

## [1] 0.7800556 0.5038275 0.1468234 0.3315217 0.9605856 0.6131944 0.8001003
## [8] 0.9673124 0.9189685 0.9553165