Multi-attribute value function sensitivity analysis

MAVF_sensitivity computes summary statistics for multi-attribute value scores of x and y given a range of swing weights for each attribute

MAVF_sensitivity(data, x, y, x_wt_min, x_wt_max, y_wt_min, y_wt_max)

Arguments

data
A data frame

x
Variable from data frame to represent x attribute values

y
Variable from data frame to represent y attribute values

x_wt_min
Lower bound anchor point for x attribute swing weight

x_wt_max
Upper bound anchor point for x attribute swing weight

y_wt_min
Lower bound anchor point for y attribute swing weight

y_wt_max
Upper bound anchor point for y attribute swing weight

Value

A data frame with added variables consisting of sensitivity analysis summary statistics for each product or service (row).

Details

The sensitivity analysis performs a Monte Carlo simulation with 1000 trials for each product or service (row). Each trial randomly selects a weight from a uniform distribution between the lower and upper bound weight parameters and calculates the mult-attribute utility score. From these trials, summary statistics for each product or service (row) are calculated and reported for the final output.

See also

MAVF_score for computing the multi-attribute value score of x and y given their respective weights SAVF_score for computing the exponential single attribute value score

Examples

# Given the following data frame that contains \code{x} and \code{y} attribute
# values for each product or service contract, we can compute how the range of
# swing weights for each \code{x} and \code{y} attribute influences the multi-
# attribute value score.

df <- data.frame(contract = 1:10,
                 x_attribute = c(0.92, 0.79, 1.00, 0.39, 0.68, 0.55, 0.73, 0.76, 1.00, 0.74),
                 y_attribute = c(0.52, 0.19, 0.62, 1.00, 0.55, 0.52, 0.53, 0.46, 0.61, 0.84))

MAVF_sensitivity(df, x_attribute, y_attribute, .55, .75, .25, .45)

#>    contract x_attribute y_attribute  MAVF_Min MAVF_1st_Q MAVF_Median MAVF_Mean
#> 1         1        0.92        0.52 0.7318977  0.7572461   0.7810366 0.7805395
#> 2         2        0.79        0.19 0.5122333  0.5474166   0.5814940 0.5807786
#> 3         3        1.00        0.62 0.8290079  0.8476105   0.8681758 0.8674587
#> 4         4        0.39        1.00 0.5426169  0.5727468   0.6031242 0.6035934
#> 5         5        0.68        0.55 0.5871887  0.6195961   0.6351362 0.6348964
#> 6         6        0.55        0.52 0.4908786  0.5256523   0.5398870 0.5398545
#> 7         7        0.73        0.53 0.6121045  0.6432748   0.6610088 0.6604361
#> 8         8        0.76        0.46 0.6034820  0.6346506   0.6559554 0.6555124
#> 9         9        1.00        0.61 0.8245081  0.8436003   0.8647067 0.8639708
#> 10       10        0.74        0.84 0.7424174  0.7645705   0.7747950 0.7751764
#>    MAVF_3rd_Q  MAVF_Max MAVF_Range
#> 1   0.8029579 0.8261519 0.09425420
#> 2   0.6132971 0.6458675 0.13363418
#> 3   0.8868259 0.9049736 0.07596571
#> 4   0.6328802 0.6640672 0.12145036
#> 5   0.6505182 0.6807529 0.09356420
#> 6   0.5550466 0.5873719 0.09649327
#> 7   0.6777540 0.7066007 0.09449621
#> 8   0.6760822 0.7048483 0.10136631
#> 9   0.8838476 0.9024729 0.07796480
#> 10  0.7860225 0.8074724 0.06505504